+++ src/allotropy/parsers/beckman_vi_cell_blu/vi_cell_blu_parser.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
 
+from dataclasses import dataclass
 from enum import Enum
-from typing import Any, NamedTuple, Optional
+from typing import Any, Optional
 
 import pandas as pd
 
@@ -62,33 +63,29 @@ class SampleProperty(Enum):
         self.data_type: Any = data_type
 
 
-class _Sample(NamedTuple):
+@dataclass(frozen=True)
+class _Sample:
     data_frame: pd.DataFrame
     row: int
 
+    def get_value(self, column: str) -> Optional[Any]:
+        if column not in self.data_frame.columns:
+            return None
+        return self.data_frame[column][self.row]
 
-def _get_value(sample: _Sample, column: str) -> Optional[Any]:
-    data_frame, row = sample
-    if column not in data_frame.columns:
-        return None
-    return data_frame[column][row]
+    def get_value_not_none(self, column: str) -> Any:
+        value = self.get_value(column)
+        if value is None:
+            msg = f"Missing value for column '{column}'."
+            raise AllotropeConversionError(msg)
+        return value
 
-
-def _get_value_not_none(sample: _Sample, column: str) -> Any:
-    value = _get_value(sample, column)
-    if value is None:
-        msg = f"Missing value for column '{column}'."
-        raise AllotropeConversionError(msg)
-    return value
-
-
-# TODO(brian): make this a method on _Sample (which then probably shouldn't be a NamedTuple)
-def get_property_from_sample(sample: _Sample, sample_property: SampleProperty) -> Any:
-    return (
-        sample_property.data_type(value=value)
-        if (value := _get_value(sample, sample_property.column_name))
-        else None
-    )
+    def get_property_value(self, sample_property: SampleProperty) -> Any:
+        return (
+            sample_property.data_type(value=value)
+            if (value := self.get_value(sample_property.column_name))
+            else None
+        )
 
 
 class ViCellBluParser(VendorParser):
@@ -119,22 +116,22 @@ class ViCellBluParser(VendorParser):
         return [
             self._get_cell_counting_document_item(_Sample(data, i))
             for i in range(len(data.index))
-            if _get_value(_Sample(data, i), "Cell count")
+            if (_Sample(data, i).get_value("Cell count"))
         ]
 
     def _get_cell_counting_document_item(
         self, sample: _Sample
     ) -> CellCountingDocumentItem:
         return CellCountingDocumentItem(
-            analyst=_get_value(sample, "Analysis by") or DEFAULT_ANALYST,
+            analyst=sample.get_value("Analysis by") or DEFAULT_ANALYST,
             measurement_aggregate_document=MeasurementAggregateDocument(
                 measurement_document=[
                     CellCountingDetectorMeasurementDocumentItem(
                         measurement_time=self._get_date_time(
-                            _get_value_not_none(sample, "Analysis date/time")
+                            sample.get_value_not_none("Analysis date/time")
                         ),
                         measurement_identifier=random_uuid_str(),
-                        sample_document=SampleDocument(sample_identifier=_get_value(sample, "Sample ID")),  # type: ignore[arg-type]
+                        sample_document=SampleDocument(sample_identifier=sample.get_value("Sample ID")),  # type: ignore[arg-type]
                         device_control_aggregate_document=CellCountingDetectorDeviceControlAggregateDocument(
                             device_control_document=[
                                 DeviceControlDocumentItemModel(
@@ -147,46 +144,45 @@ class ViCellBluParser(VendorParser):
                             processed_data_document=[
                                 ProcessedDataDocumentItem(
                                     data_processing_document=DataProcessingDocument(
-                                        cell_type_processing_method=_get_value(
-                                            sample, "Cell type"
+                                        cell_type_processing_method=sample.get_value(
+                                            "Cell type"
                                         ),
-                                        minimum_cell_diameter_setting=get_property_from_sample(
-                                            sample, SampleProperty.MINIMUM_DIAMETER
+                                        minimum_cell_diameter_setting=sample.get_property_value(
+                                            SampleProperty.MINIMUM_DIAMETER
                                         ),
-                                        maximum_cell_diameter_setting=get_property_from_sample(
-                                            sample, SampleProperty.MAXIMUM_DIAMETER
+                                        maximum_cell_diameter_setting=sample.get_property_value(
+                                            SampleProperty.MAXIMUM_DIAMETER
                                         ),
-                                        cell_density_dilution_factor=get_property_from_sample(
-                                            sample, SampleProperty.DILUTION
+                                        cell_density_dilution_factor=sample.get_property_value(
+                                            SampleProperty.DILUTION
                                         ),
                                     ),
-                                    viability__cell_counter_=get_property_from_sample(
-                                        sample, SampleProperty.VIABILITY
+                                    viability__cell_counter_=sample.get_property_value(
+                                        SampleProperty.VIABILITY
                                     ),
-                                    viable_cell_density__cell_counter_=get_property_from_sample(
-                                        sample, SampleProperty.VIABLE_CELLS_ML
+                                    viable_cell_density__cell_counter_=sample.get_property_value(
+                                        SampleProperty.VIABLE_CELLS_ML
                                     ),
-                                    total_cell_count=get_property_from_sample(
-                                        sample, SampleProperty.CELL_COUNT
+                                    total_cell_count=sample.get_property_value(
+                                        SampleProperty.CELL_COUNT
                                     ),
-                                    total_cell_density__cell_counter_=get_property_from_sample(
-                                        sample, SampleProperty.TOTAL_CELLS_ML
+                                    total_cell_density__cell_counter_=sample.get_property_value(
+                                        SampleProperty.TOTAL_CELLS_ML
                                     ),
-                                    average_total_cell_diameter=get_property_from_sample(
-                                        sample, SampleProperty.AVERAGE_DIAMETER
+                                    average_total_cell_diameter=sample.get_property_value(
+                                        SampleProperty.AVERAGE_DIAMETER
                                     ),
-                                    average_live_cell_diameter__cell_counter_=get_property_from_sample(
-                                        sample, SampleProperty.AVERAGE_VIABLE_DIAMETER
+                                    average_live_cell_diameter__cell_counter_=sample.get_property_value(
+                                        SampleProperty.AVERAGE_VIABLE_DIAMETER
                                     ),
-                                    viable_cell_count=get_property_from_sample(
-                                        sample, SampleProperty.VIABLE_CELLS
+                                    viable_cell_count=sample.get_property_value(
+                                        SampleProperty.VIABLE_CELLS
                                     ),
-                                    average_total_cell_circularity=get_property_from_sample(
-                                        sample, SampleProperty.AVERAGE_CIRCULARITY
+                                    average_total_cell_circularity=sample.get_property_value(
+                                        SampleProperty.AVERAGE_CIRCULARITY
                                     ),
-                                    average_viable_cell_circularity=get_property_from_sample(
-                                        sample,
-                                        SampleProperty.AVERAGE_VIABLE_CIRCULARITY,
+                                    average_viable_cell_circularity=sample.get_property_value(
+                                        SampleProperty.AVERAGE_VIABLE_CIRCULARITY
                                     ),
                                 ),
                             ]